+++ src/lib.rs
@@ -263,6 +263,7 @@ pub mod oversample;
 pub mod pan;
 pub mod pattern;
 pub mod prelude;
+pub mod realnet;
 pub mod resample;
 pub mod scale;
 pub mod sequencer;